Workaround when Carp is blacklisted
This is pretty ugly but acceptable as we
are not using Carp, and seems that the only
problem comes from FileHandle using IO::Seekable.

If the customer explicitely blacklist Carp
from being compiled in we can temporary load it.

Need to be sure we are not saving it.

view xtestc/0235.t