declare_static_sections memory blocks
This is avoiding free on static elements,
most of the existing patches to struct can go away
as this is providing an extra layer on top of
safesysrealloc & safesysfree.

Only free memory chunk not declared as static,
and do a malloc + copy for memory during realloc
when the memory is static.

Easy way to test is to add something like this to the
end of declare_static_sections:
---
Safefree( &sv_list[1] );
{
    SV *foo = saferealloc( &sv_list[1], 2 * sizeof( sv_list[1] ) );
    Safefree( foo );
}