meta shared_he - step 1
Attempt to use meta for shared_he

All the changes on the pm side are done,
but this is going to add a lot of complexity
to the init_pl_strtab and init_assign functions
which are still using the obsolete sharedhe_list.

Not sure it's a good idea to pursue this effort.