Static malloc declarations
Use a malloc section to declare precisely all
the malloc we are going to use so no need to
call a declare function anymore or using a slow
HV internal mechanism.

Note that we are not using 'sizeof' anymore but some
hardcoded values. Need to autocheck them at startup
to be sure everything is ok.