Simplify the EV::API patch
allow EV bootstrap by removing SVf_READONLY
from EV::API UV.

This is a simplified version of what we were used to
do in bc524 with something like this before calling
bootstrap. And this do not need to happen at run time
but can simply be adjusted at compile time.

{
    GV *sym;
    sym = gv_fetchpv("EV::API", 0, SVt_PV);
    if (sym && GvSVn(sym)) SvREADONLY_off(GvSVn(sym));
}