bootstrap B with workaround for constants
Avoid duplicate B:: constants:
this is a monkey patch to do not save B:: constants
as we are going to bootstrap them when required.

Add some extra B cleanup.
One idea could be to detect if the compiled program
requires B and if not blacklist it as it previously was.

Note that Moose GH #55 is now passing.