B can be stripped on demand
This is a first move to GH #71,
when the customer provides -UB then
the previous behavior is restored and B
is stripped from the binary.

In the future perlcc would be smarter and detect
when B is required or not and provides itself the -UB
under the hood, leaving then the customer the -uB flag
to force to enable it to fix some false positives.