Avoid deep recursion calls using tail calls
This is only a problem when compiling large programs,
no unit test from the testsuite currently show
that issue.

This is improving the deep recursion protection which
was removed as part of commit 834b4f34c87.
Such a protection is required when compiling large programs.

Note: can also consider using Sub::Call::Tail but try
to preserve the B::C dependencies list as small as possible.