Do not save CVs/GVs for CHECK, BEGIN & UNITCHECK blocks
We should not save CVs for BEGIN, CHECK and UNITCHECK
The GV is simply ignored if no AV, HV, format, IO is used.
( @BEGIN ) If one of them is used then we are skipping
the CV.

Note that anything pointing to a backref to the CV or GV will
segfault. We do not expect this to be common.
i.e.: *main::BEGIN or &main::BEGIN