Tweak Dockerfile & Jenkinsfile to run Perl 528
This is avoiding the 'chicken & egg' issue where
we cannot start working on B::C before having the
full Perl RPM stack available.

All we need is the perl RPM itself and a few extra
modules. Provide a simple script src/install-perl-modules.pl
which will rely on App::cpanminus to install dependencies
if needed. (note we are not using a cpanfile as we also
need to deal with a few exceptions)