Fixup eval_pv to open DATA from unicode namespace
Case BC-2142:

We are adding one dirty 'eval_pv' trick to open the DATA
section. But when the DATA section becomes to one utf8 namespace
we should let the open be aware of it.

This is fixing the final tests of uni/method.t

This can be reproduced using the following simple program.
_______________________
use utf8;
package ʑyourmom;
no strict 'refs';

print qq[done\n];

__DATA__
Some data