Provide a custom mode to preset every OPs
The lazy OP mode is a good thing for most scripts,
as we are barely using all codepath, and this providing
a significant memory reduction and memory speed.

But in some specific cases, like daemons forking,
we will slow down every child process as any  of them
would have to preload every OPs.

This commit is providing a custom compilation mode which
can be enabled by setting

    BEGIN { *B::C::SETUP_ALL_OPS = sub { 1 } }

This is useful for a few daemons. Note that this has to
be set at compile time, and B::C is throwing away *B::C::*
So this would not be saved.

Add a single unit test to check this mode.
We could also consider running the whole test suite with this
custom mode.