Support for Lazy Regex
Use a custom 'LazyRegExp' struct to store
data to build the RegExp object itself
    - a pointer to the RegExp object
    - the regex string
    - flags
    - refcnt

Note that a RegExp can be shared so we pay
only once the compilation cost.

Inside a pmop_list entry we are stealing the pmflags
to store our `position` in the rx_list so we can retrieve
all data.

This position is shifted using `PMf_BASE_SHIFT` to avoid being
seen ourself as a PMf. (maybe can steal another integer value)

We are then compiling the RegExp the very first time we see it
and reuse it in all other cases.