Use pmopaux to store the index
We cannot steal the REGEXP pointer to store our index.
Messing with the flags can also have some limitations and
we can raise some unpredictable bugs.

This commit is storing the index to access to the rx_list
in its own struct named PMOPAUX.

We could also have considered adding this index at the end
of the PMOP struct itself.