Use SvREFCNT to increase directly the RefCnt
Avoid using a for loop to increase the RegExp refcnt
and use `SvREFCNT` helper.

Note that this is also fixing a bug for refcnt was starting at 1.
It should start at 0, as we are adding that counter to the existing
RefCnt state (which should be 1).

We could also use a `SvREFCNT(rx.regex) = rx.refcnt` and start at 1.
But this is taking a risk if for any reasons the RefCnt is higher than one.