Add support for Sub::Defer
Defer all Sub::Defer subs at compile time

Moo is using Sub::Defer to generate functions
on their first call, so all the eval are not
run at compile time.

B::C does not care about the compile time,
and as the global cache is using CODE references
as key, this is a bad idea to defer functions.